backend/django_server/requirements.txt
bcrypt==3.1.7
cryptography==2.8
Django==3.0.2
django-cors-headers==3.11.0
django-extensions==3.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
evdev==1.3.0
Faker==11.3.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
mysqlclient==2.1.0
mysql-connector-python==8.0.26
pipenv==11.9.0
PyJWT==1.7.1
python-dateutil==2.7.3
python-dotenv==0.10.3
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
tabulate==0.8.9
testresources==2.0.1
text-unidecode==1.3
ua-parser==0.10.0
user-agents==2.2.0
uuid
virtualenv==20.0.17
virtualenv-clone==0.3.0